requirements.txt
# AI Novel Media Agent 依赖包
# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库和缓存
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jwt==3.3.4
cryptography==41.0.7

# 文件处理
minio==7.1.16
Pillow==10.1.0
python-magic==0.4.27

# 视频处理
moviepy==1.0.3
opencv-python==4.8.1.78
pydub==0.25.1

# AI和LLM
openai==1.3.5
anthropic==0.7.7
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99

# 网络和API
httpx==0.25.1
websockets==12.0
aiohttp==3.9.1
requests==2.31.0

# 数据处理
pandas==2.1.3
numpy==1.24.4
scikit-learn==1.3.2

# 工具和工具
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
click==8.1.7
rich==13.7.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# 部署和监控
gunicorn==21.2.0
supervisor==4.2.5
prometheus-client==0.19.0

# 其他
paramiko==3.4.0
python-dateutil==2.8.2
tzdata==2023.3
colorama==0.4.6